move LinearAlgebra and Dates to package extensions
