+++ src/Compat.jl
@@ -1,10 +1,5 @@
 module Compat
 
-import Dates
-using Dates: Period, CompoundPeriod
-
-import LinearAlgebra
-
 include("compatmacro.jl")
 
 # NOTE these `@inline` and `@noinline` definitions overwrite the definitions implicitly
@@ -236,13 +231,6 @@ else
     using Base: @constprop
 end
 
-# https://github.com/JuliaLang/julia/pull/40803
-if VERSION < v"1.8.0-DEV.300"
-    function Base.convert(::Type{T}, x::CompoundPeriod) where T<:Period
-        return isconcretetype(T) ? sum(T, x.periods) : throw(MethodError(convert, (T, x)))
-    end
-end
-
 # https://github.com/JuliaLang/julia/pull/39245
 if VERSION < v"1.8.0-DEV.487"
     export eachsplit
@@ -327,17 +315,6 @@ if VERSION < v"1.8.0-DEV.1494" # 98e60ffb11ee431e462b092b48a31a1204bd263d
     allequal(itr) = isempty(itr) ? true : all(isequal(first(itr)), itr)
     allequal(c::Union{AbstractSet,AbstractDict}) = length(c) <= 1
     allequal(r::AbstractRange) = iszero(step(r)) || length(r) <= 1
-end
-
-# https://github.com/JuliaLang/julia/commit/bdf9ead91e5a8dfd91643a17c1626032faada329
-if VERSION < v"1.8.0-DEV.1109"
-    # we do not add the methods for == and isless that are included in the above
-    # commit, since they are already present in earlier versions.
-    import Base: /, rem, mod, lcm, gcd, div
-    for op in (:/, :rem, :mod, :lcm, :gcd)
-        @eval ($op)(x::Period, y::Period) = ($op)(promote(x, y)...)
-    end
-    div(x::Period, y::Period, r::RoundingMode) = div(promote(x, y)..., r)
 end
 
 # This function is available as of Julia 1.7.
@@ -640,4 +617,8 @@ end
 
 include("deprecated.jl")
 
+if !isdefined(Base, :get_extension)
+    include("../ext/CompatLinearAlgebraExt.jl")
+    include("../ext/CompatDatesExt.jl")
+end
 end # module Compat
